Reorganize optimize-images, reports, and shadows into site status
