--- style/optimize-images.css
@@ -1,3 +1,0 @@
-/**
- * Styles for optimize-images tool.
- */
