--- style/reports.css
@@ -1,3 +1,0 @@
-/**
- * Styles for reports module.
- */
